Migration and model changes for CorporateInfoPages.
Change CorporateInformationPage to be an Edition. As a general principle, trying
to avoid adding fields to the Edition base model: so far the exception is
corporate_information_page_type_id, which follows the precedent of
publication_type_id to point to the fake model.

Slugs in CIPs are meaningless, as they depend on the CIPType and are only
unique within each organisation. As a consequence, the migration does not
create a document slug at all, relying on the default behaviour of using the
db id.

Previously CIP had a polymorphic belongs_to relationship to organisation and
worldwide_organisation. Here we reuse the existing ::Organisation and
::WorldwideOrganisation mixins, which are actually many-to-many, but we add
validation to ensure that only one organisation is allowed overall.